requirements.txt
psycopg2-binary
python-dotenv
fer[ALL]
tensorflow